+++ CHANGELOG.md
@@ -13,12 +13,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security
 
+## [1.8.0]
+
+### Added
+### Changed
+    - Issue 329 - Fix docs so they deploy when we make a new release
+    - Issue 321 - Make SWOT Version D the default collection, and update docs
+    - Issue 292 - Ensure collection_name query parameter returns SWOT data for child collections
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## [1.7.0]
 
 ### Added
     - Issue 304 - Add wse_sm, wse_sm_u, wse_sm_q, wse_sm_q_b (smoothed WSE) fields for Version D Node features
     - Issue 261 - Add qual_f_b for Version D Lake features
-    - Added Regression Test Suite to test live API in UAT and OPS 
+    - Added Regression Test Suite to test live API in UAT and OPS
     - Issue 306 - CORS header missing on responses with 4xx error results
 ### Changed
     - Updated Python to 3.12